chore: consolidate game systems into single directory
