+++ src/game/systems/loot.js
@@ -1,5 +1,7 @@
 // WEAPONS-INTEGRATION: basic loot table rolling
 import { LOOT_TABLES } from '../../data/lootTables.js';
+import { rollWeaponDropForZone } from '../../data/lootTables.weapons.js';
+import { ZONES } from '../../data/zones.js';
 
 export function toLootTableKey(id = '') {
   return (id || '')
@@ -18,4 +20,16 @@ export function rollLoot(key, rng = Math.random) {
     if (roll <= acc) return entry.item;
   }
   return null;
+}
+
+export function onEnemyDefeated(state) {
+  const zoneKey = state.world?.zoneKey ?? ZONES.STARTING;
+
+  const weaponDrop = rollWeaponDropForZone(zoneKey);
+  if (weaponDrop) {
+    state.inventory.items.push(weaponDrop);
+    state.log.push(`You found: ${weaponDrop.name}.`);
+  }
+
+  // … add other rewards (xp, coins) as you already do
 }
